+++ src/utils/persistentStore.ts
@@ -9,6 +9,7 @@ import Store from 'electron-store';
 import { SerialPortOpenOptions } from 'serialport';
 import { v4 as uuid } from 'uuid';
 
+import { inMain as safeStorage } from '../../ipc/safeStorage';
 import logger from '../logging';
 import packageJson from './packageJson';
 
@@ -44,6 +45,31 @@ export const persistIsFavorite = (serialNumber: string, value: boolean) =>
     sharedStore.set(`${serialNumber}.fav`, value);
 export const getPersistedIsFavorite = (serialNumber: string) =>
     sharedStore.get(`${serialNumber}.fav`, false) as boolean;
+
+export const persistApiKey = async (keyName: string, apiKey: string) => {
+    const canEncrypt = await safeStorage.isEncryptionAvailable();
+    if (!canEncrypt) {
+        throw new Error('Encryption not available');
+    }
+
+    const encrypted = await safeStorage.encryptString(apiKey);
+    sharedStore.set(`apiKeys.${keyName}`, Array.from(encrypted));
+};
+export const getPersistedApiKey = async (keyName: string) => {
+    const canDecrypt = await safeStorage.isEncryptionAvailable();
+    if (!canDecrypt) {
+        throw new Error('Decryption not available');
+    }
+
+    const encrypted = sharedStore.get(
+        `apiKeys.${keyName}`,
+        {} as Array<number>
+    );
+
+    return encrypted.length > 0
+        ? safeStorage.decryptString(Buffer.from(encrypted))
+        : '';
+};
 
 export const persistSerialPortSettings = (
     serialNumber: string,